Move helper to utils package Add comments
